Fixes cache and directory problems
Cache was not cleaned after "ng" installation, so we make `npm cache clean --force` as soon as possible to not include it in image.

`ng new` doesn't want to install in current directory, the better I can do is to keep the current directory, step up one directory, install project and return back in the project directory. Don't know why that worked before.